clinical_data_explorer_simple/requirements.txt

streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
duckdb==0.9.2
pyreadstat==1.2.6
scipy==1.11.4
statsmodels==0.14.1
plotly==5.18.0
openpyxl==3.1.2
python-docx==1.1.0
PyYAML==6.0.1
Jinja2==3.1.3
lifelines==0.27.8
pyarrow==14.0.2
